Estimate coefficients of a high-dimensional GLM

This is the generic function and methods associated with the adjusted MLE for high dimensional generalized linear models (glm).

adjust_glm(glm_output, verbose = FALSE, echo = TRUE, ...)

Arguments

glm_output
An object from the class glm, for example the output from the glm function

verbose
If TRUE, print progress at every step.

echo
If TRUE, returns the original glm_output.

...
Additional arguments.

Details

The function adjust_glm is built on the glm function: the input should be from the class glm, such as the output from glm. Currently, it works with binary regressions only, for which it calls the function adjust_binary to estimate coefficients and their standard deviations. In high dimensions, the magnitude of MLE is biased upward, and the variance estimates from classical theory based on inverse Fisher information underestimates the true variability. adjust_binary adjusts the MLE and variance estimates applying the asymptotic theory assuming \(p, n\to\infty\) (\(p<n\)), and is exact when covariates \(X\) is multivariate Gaussian.

adjust_glm creates an object from class glmadj. summary.glmadj creates a data table of adjusted coefficients, estimates of the standard error and p-values using a two-sided t-test. If the covariates are multivariate Gaussian, the p-values are asymptotically from a uniform distribution.

predict.glmadj computes the prediction \(X^\top \hat{\beta}^{\mathrm{Adj}}\) for new input data. The new data should be of the same format which you used to fit the glm. If there's no new data, it outputs the estimated \(X^\top \hat{\beta}_{\mathrm{Adj}}\) for the original dataset.

See also

Examples

# Problem size
n <- 1000L
p <- 300L
# Generate data matrix
X <- matrix(rnorm(n*p, 0, 1), n, p)
# Generate parameters
beta <- rep(c(0, 1), p / 2) / sqrt(p) * 2
#' ## No intercept model
# Sample response vector
Y <- rbinom(n, 1, 1 / (1 + exp(- X %*% beta)))
# Fit a glm, by default computes a logistic regression
fit <- glm(Y ~ X + 0, family = binomial, x = TRUE, y = TRUE)
adjusted_fit <- adjust_glm(fit)
# Print summary
print(summary(adjusted_fit), max = 40)

#> Call:
#> adjust_glm(glm_output = fit)
#> 
#>      adjusted_mle         std z.value   p.value    
#> X1    -0.05120675  0.09129046 -0.5609 0.5748513    
#> X2     0.19716131  0.09126411  2.1603 0.0307465 *  
#> X3     0.01883666  0.09350460  0.2015 0.8403454    
#> X4     0.23230363  0.09223395  2.5186 0.0117811 *  
#> X5     0.06361850  0.09457172  0.6727 0.5011375    
#> X6     0.14050059  0.09135025  1.5380 0.1240382    
#> X7     0.11987165  0.09263620  1.2940 0.1956639    
#> X8     0.13572113  0.09393962  1.4448 0.1485226    
#>  [ reached getOption("max.print") -- omitted 292 rows ]
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1

# Predict on new data
head(predict(adjusted_fit))

#> [1] -0.47390481 -0.89463130  0.39557465 -1.96782805 -0.04959725 -0.49277543

# Extract adjusted coefficients
head(adjusted_fit$coef_adj)

#>          X1          X2          X3          X4          X5          X6 
#> -0.05120675  0.19716131  0.01883666  0.23230363  0.06361850  0.14050059 

# Extract adjusted standard error
head(adjusted_fit$std_adj)

#>        X1        X2        X3        X4        X5        X6 
#> 0.1644027 0.1643552 0.1683900 0.1661018 0.1703118 0.1645103